HLS-24 Allow audio track to be ignored with the `-n` commandline argument
This forces `-an` to be passed to `ffmpeg`.

It's not generally required, as missing audio tracks will be handled by default, but this allows the user to address inputs where `ffmpeg` reports the Audio type as `none`:

    Stream #0:0: Video: flv1, 1 reference frame, yuv420p, 1280x720, 199 kb/s, 8.33 tbr, 1k tbn, 1k tbc
    Stream #0:1: Audio: none, 0 channels

This test stream was the result of a webcam being published into an RTMP relay with FFMPEG and then pulled back out the other side.